Resolve example.com instead of localhost
The IP addresses of localhost. heavily depend on the configuration of
the local resolver, producing unreproductible and unexpected
consequences when running the test suite on Travis-CI.

On the other hand, the IPv4 and IPv6 addresses of example.com seems to
be somewhat stable.

Update the test suite to check for example.com addresses when testing
names resolution.