Add a function to query SRV records to build rules
Add a srv() function that queries the DNS for SRV records and return
endpoints usable in from / to host rules.

This makes it easier to maintain ruleset with dynamic resources by
postponing the resolution of these SRV records the same way name
resolution only happen when gererating the final final firewall rules.