Improve abstraction of hosts
A variable is basically a host or a port, and a host can be either an IP
address, a hostname, another variable, and so on.

Use our abstraction of host_list_item outside of lists when a single
host is expected, as a variable value, or in a rule.  This allows us to
have a better abstraction.